Make CQLStatement.validate() non-blocking by specification
Unless I've missed something, CQLStatement.validate() never blocks
(contrarily to checkAccess), it just does basic checks (in fact, it
looks like it's only used by DDL statements and thus all validation
could be moved to prepare time and the method eliminated, but it's a
different story). Also feels unlikely we'd screw this by mistake in
the future.

So specify that this should be the case, but otherwise
slightly simplify the QueryProcessor code on that assumption.