Better fix for LIMIT 0 queries
I was surprised by the special casing of 'LIMIT 0' in
ReadCommand.executeLocally() since it's such a useless case that
optimizing for it feels more like code pollution than anything else (and
if we really want to optimize, we should do so coordinator side).

Turns out this was added because the code was broken otherwise, which
surprised me, and that was because DataLimits.Counter was overriding
one of StoppingTransformation.attachTo method, but not the new one added
for PartitionPublisher. And that feels like a genuine bug (though I'm
not 100% sure why it didn't seem to have more visible consequence at
this point) so the commit fixes that. And hence remove the 'LIMIT 0'
for the reason mentioned above.