APOLLO-909
make Domains, PartitionedEnumSet, ParmissionSets more efficient

move permission checks into QueryState

Add ANONYMOUS_ROLE
add negative caching, better null checks

consolidate authorization stuff, eliminate JMXPermissionsCache,
clueter wide roles invalidation

integrate credentials-cache into roles-cache

`QueryState` gets the ability to "inject" resource-permissions. This is
needed for Row-Level-Access-Control (RLAC), as `QueryState` is meant to
contain all resource-permissions for all roles a user is assigned to.

`AuthenticatedUser´ gets a new method `getLoginRole()` that returns the
authenticated user's role. This change is for DSE proxy authentication.

`IRoleManager` gets a new method `transitiveRoleLogin()` that indicates
whether all transitive roles should be considered to allow a user
to login, which is required for LDAP authentication, or not, which
is the default.

Simplify CassandraAuthorizer::permissionsForRole. The resource is
guaranty to be always different as it is the clustering key and that we
are restricting on the partition key (role)

Removes the cache negative. If a null value was returned by the loader
it would have been tiggered a ClassCastException. I believe that it is
up to the loader to always return a proper value.

Make Role really immutable

Make Auth asynchronous

Remove PermissionSet ConcurrentMap and reduce instantiations when
checking permissions in UserRolesAndPermissions

Fix loading of Prepared statements