Re-enable the NodeSync service by default
APOLLO-1091 has been committed and, to the best of my knowledge, this
was the last outsanding 'blocking' bug for NodeSync. So it's high time
we switch this back since this is how we want thing to work for release.

Side-note: this _only_ enable the service itself by default. But
NodeSync is still off by default on all tables, so by default the
service will do nothing up until the user enable NodeSync on a table.