Rename WatcherThread, and increase watcher default sleep time to 50us
To avoid confusion with the JVM Runtime WatcherThread.
We rely on parkNanos(1) to sleep for roughly 50us, when in fact the
behaviour will vary based on OS/JVM. This way we better express our
wishes. This issue manifested as high CPU utilization on Ubuntu 17.10
so kernel 4.13. It has also been known to happen on RH distributions
and MacOS(I think).