Authentication & Authorization improviements
patch by Robert Stupp & Benjamin Lerer; reviewed by Benjamin Lerer & Robert Stupp for APOLLO-909

make Domains, PartitionedEnumSet, ParmissionSets more efficient
move permission checks into QueryState

consolidate authorization stuff, eliminate JMXPermissionsCache, clueter wide roles invalidation

integrate credentials-cache into roles-cache

`QueryState` gets the ability to "inject" resource-permissions. This is
needed for Row-Level-Access-Control (RLAC), as `QueryState` is meant to
contain all resource-permissions for all roles a user is assigned to.

`AuthenticatedUser´ gets a new method `getLoginRole()` that returns the
authenticated user's role. This change is for DSE proxy authentication.

`IRoleManager` gets a new method `transitiveRoleLogin()` that indicates
whether all transitive roles should be considered to allow a user
to login, which is required for LDAP authentication, or not, which
is the default.

Simplify CassandraAuthorizer::permissionsForRole. The resource is
guaranteed to be always different as it is the clustering key and that we
are restricting on the partition key (role)

Make Role immutable

Make Auth asynchronous