Make thread local fields static everywhere, make singleton ref fields static too.
For thread local the expected usage is as static final, anything else
should be done with care and carefully reasoned about. The cases
converted had no reason to violate this best practice.

Furthermore, while the JVM can "reason through" that final fields on a
constant reference (like a singleton) are in fact also constant the
same is not always true for reference fields. Even reasoning through
primitive finals is not a given. It follows that final fields on
singletons should also be static (at least where it is not hard to do
so).