More accurate exception message on CAS contention timeout
If there are multiple parties contending for a CAS on the same
partition, they may easily timeout due to reaching
cas_contention_timeout. In this case the previous general
WriteTimeoutException message was a bit misleading, because it could
make the user think there was a performance- or overload-related
problem and not concurrency-related one.

This commit leaves the type of the exception the same, but the message
describes the nature of the timeout more accurately.

Patch by Piotr Kołaczkowski, reviewed by Alex Petrov for DB-812