STAR-258: Fix Scrubber
It fixes the problem introduced in one of the previous tickets, namely STAR-15 or something aroung.

seekToNext row is expected to position dataFile at the data position retrieved from the index entry. This is where the buggy impl finished. However, in the original implementation, we continue iterating over the index until we get the data position to which we can seek data file properly.