STAR-13 Add the Unified Compaction Strategy.
This is the implementation of UnifiedCompactionStrategy, whish is
intended to not only replace all other compaction strategies and CSM,
but also to optimally choose the configuration that will result in the
minimum read and write costs, given a specific workload and dataset size.

The strategy will choose either leveled or tiered merge policies
depending on the workload and the costs associated with user queries and
inserts.

This strategy should be considered experimental at this stage.

--

This patch also introduces a compaction simulation that can be run with:

ant compactionSim

With cmd line arguments:

ant compactionSim--args="-wl R50_W50 -t adaptive"

Co-authored-by: Stefania Alborghetti <stefania.alborghetti@datastax.com>
Co-authored-by: Branimir Lambov <branimir.lambov@datastax.com>