STAR-254 Port DSE custom CQL types (#107)
* STAR-254 copy DSE 6.8.11 additional types

* STAR-254 change DataStax copyrights to Apache

* STAR-254 add esri-geometry-api

2.2.4 is currently the latest release.

This is a different version then the one in DSE 6.8.11
(1.2.1). This one doesn't have the following dependencies:
org.codehaus.jackson:jackson-core-asl and org.json:json.
Instead it relies on com.fasterxml.jackson.core:jackson-core
that is already present on C* classpath.

* STAR-254 fix DateRangeUtil

by moving some Apache Lucene™ methods. Using the exact same code
as DSE ensures compatibility and at the same avoids dragging in
the whole lucene jar to this project.

* STAR-254 adjust to OSS serializer interface

DateRangeSerializer was adjusted to the new interface.
The instances are created without slicing a new ByteBuffer,
unfortunately it implied manual offset counting (duh!).

Geometric TypeSerializer uses a sliced buffer as using the
ValueAccessor is not possible in this case (the buffer is
eventually passed to an external library).

* STAR-254 fix DateRangeSerializerTest

Adjust to Cassandra test parametrization.

* STAR-254 fix tests

* STAR-254 fix GeometryIntegrationTest

Allocation types parameters are removed, a dedicated tests
for endianess are added instead.

Fake (!) endianess logic for OGCGeometry was simplified.

* STAR-254 fix geo tests by adjusting esri

Polygon assertion was changed for the equivalent polygon.
Esri is now more consistent in results of to* methods
(polygon points are always listed clockwise).