STAR-13 Implement compaction shards
The unified compaction strategy is now decoupled from
CompactionStrategyManager by implementing its own thin compaction
container.

When a compaction candidate needs to be produced, the strategy takes a
snapshot of the eligible sstables and disk boundaries. It then applies
a set of equivalence classes, which implement the same partitioning of
sstables currently performed by CSM.

One of the equivalence classes splits sstables across token range
boundaries, or shards. The number of shards is specified in the
compaction properties, and the shards are created by splitting the
local ranges by this number. sstables are assigned to a shard by
looking at their first partition key.

When flushing or compacting, a specialized writer splits sstables
at the shard boundaries. If the current sstable is larger than a
minimum sstable size that can be specified in the compaction properties,
then it is split when a boundary is reached.

getNextBackgroundTask() now returns a list of tasks, which are processed by
the compaction executor asynchronously.

The following minor updates have been performed:

- the adaptive algorithm now searches for better choices of W every 5 minutes
  rather than 2 minutes;
- the cost calculator now uses a read multiplier of 0.1 rather than 0.25;
- all sstables in an bucket are compacted if their number is >= T. Compactions
  no longer stop at T or F. This may skip levels but has proven very effective
  in tests when switching from tiered to levelled.

The documentation for the unified strategy has been added as a mark down
document.

Reviewed by Branimir Lambov, Dimitar Dimitrov, Justin Chu, Aleksandr Sorokoumov

Co-authored-by: Stefania Alborghetti <stefania.alborghetti@datastax.com>
Co-authored-by: Branimir Lambov <branimir.lambov@datastax.com>